MSI-886: Need to move ReservationBuilderInterface to ReservationApi module.
