+++ app/code/Magento/InventoryShipping/composer.json
@@ -12,7 +12,6 @@
         "magento/module-inventory-source-selection-api": "*",
         "magento/module-inventory-source-selection": "*",
         "magento/module-inventory-reservations-api": "*",
-        "magento/module-inventory-reservations": "*",
         "magento/module-sales": "*",
         "magento/module-shipping": "*",
         "magento/module-ui": "*",
